TrafficSignRecAPI/requirements.txt
Flask==1.1.1
Tensorflow==2.0
Pillow
flask-cors
